SCP-3192: Actus module refactoring
* Split actus-core and Marlowe specific code
* Module renaming
* Dropped quick-check tests
