+++ marlowe-actus/actus-core/src/Actus/Model.hs
@@ -1,0 +1,15 @@
+module Actus.Model
+  ( module Actus.Model.Applicability
+  , module Actus.Model.ContractSchedule
+  , module Actus.Model.Payoff
+  , module Actus.Model.StateInitialization
+  , module Actus.Model.StateTransition
+  )
+  where
+
+import Actus.Model.Applicability
+import Actus.Model.ContractSchedule
+import Actus.Model.Payoff hiding (contractTerms, riskFactors)
+import Actus.Model.StateInitialization
+import Actus.Model.StateTransition hiding (maturity)
+
